move GridRunner into its own file so other consumers (including us!) can extend it to get some good base test functionality. also added a pretty arbitary performance safety check to make sure the grid doesnt take toooo long to render (even on bulid servers etc)
